fix propTypes validation
