+++ src/components/app/app.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import Main from '../main/main';
 import PropTypes from "prop-types";
+import {offerType} from '../../types/offer';
 
 const App = ({offers}) => {
   return (
@@ -9,7 +10,9 @@ const App = ({offers}) => {
 };
 
 App.propTypes = {
-  offers: PropTypes.array.isRequired,
+  offers: PropTypes.arrayOf(
+      offerType.isRequired
+  ).isRequired,
 };
 
 export default App;
