+++ src/data/constants.js
@@ -1,0 +1,3 @@
+const HOUSING_TYPES = [`Apartment`, `Room`, `House`, `Hotel`];
+
+export {HOUSING_TYPES};
